+++ tests/scrapers/verapdf_test.py
@@ -15,6 +15,8 @@ This module tests that:
     - For files that are valid PDF 1.7 or 1.4 but not valid PDF/A, MIME type,
       version and streams are scraped correctly but they are reported as
       not well-formed.
+    - Scraper can be run for files without PDF extension.
+    - --nonpdfext flag is handled properly.
     - The scraper supports MIME type application/pdf with versions A-1b
       when well-formedness is checked, but does not support them when
       well-formedness is not checked. The scraper also does not support made
@@ -24,7 +26,7 @@ from __future__ import unicode_literals
 
 import pytest
 
-from file_scraper.verapdf.verapdf_scraper import VerapdfScraper
+from file_scraper.verapdf.verapdf_scraper import VerapdfScraper, filter_errors
 from tests.common import (parse_results, partial_message_included)
 
 MIMETYPE = "application/pdf"
@@ -127,6 +129,16 @@ def test_scraper_no_file_extension(evaluate_scraper):
     scraper.scrape_file()
 
     evaluate_scraper(scraper, correct)
+
+
+def test_error_filtering():
+    """Test flag creation based on version"""
+    error = \
+        "Oct 17, 2022 3:32:30 PM org.verapdf.apps.utils." \
+        "ApplicationUtils filterPdfFiles\n" \
+        "SEVERE: File /mypath/--nonpdfext doesn't exist.\n" \
+        "ERROR: A real error not to be removed."
+    assert filter_errors(error) == "ERROR: A real error not to be removed."
 
 
 def test_is_supported():